菜品信息的修改

+++ sky-server/src/main/java/com/sky/mapper/SetmealDishMapper.java
@@ -15,5 +15,5 @@ public interface SetmealDishMapper {
      * @param dishIds
      * @return
      */
-    List<Long> getSetmealsByDishIds(List<Integer> dishIds);
+    List<Long> getSetmealsByDishIds(List<Long> dishIds);
 }
